Refactor: Move ResetPasswordModal logic to ResetPasswordPage and remove unused files
