+++ src/components/UI/Modal/Components/ForgotPasswordModal/ForgotPasswordModal.module.css
@@ -63,10 +63,16 @@
     border-color: var(--color-accent);
   }
   
-.error {
+  .errorInput {
+    border-color: red;
+    background-color: rgba(255, 0, 0, 0.1);
+  }
+  
+  .error {
     color: red;
     font-size: 12px;
-}
+    margin-top: 5px;
+  }
 
 button {
     width: 100%;
