--- src/components/UI/Modal/Components/ResetPasswordModal/ResetPasswordModal.module.css
@@ -1,192 +1,0 @@
-.modalOverlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.modalContent {
-  background-color: var(--color-gray);
-  padding: 20px;
-  border-radius: 20px;
-  width: 90%;
-  text-align: center;
-}
-
-.modalHeader {
-  text-align: start;
-  font-size: 18px;
-  font-weight: bold;
-  color: var(--color-main);
-  margin-bottom: 15px;
-}
-
-.modalText {
-  font-size: 14px;
-  color: #555;
-  text-align: start;
-  margin-bottom: 20px;
-}
-
-.inputField {
-  width: 100%;
-  height: 50px;
-  padding: 12px;
-  border: 1px solid rgba(47, 47, 47, 0.2);
-  border-radius: 15px;
-  font-size: 13px;
-  line-height: 18px;
-  background-color: var(--color-light);
-  transition: border-color 0.3s ease, box-shadow 0.3s ease;
-  margin-bottom: 16px;
-  padding-right: 48px;
-  color: inherit;
-}
-
-.inputField.inactive {
-  border-color: rgba(47, 47, 47, 0.2); 
-  background-color: var(--color-light);
-}
-
-.inputField:focus {
-  outline: none;
-  border-color: var(--color-accent);
-  box-shadow: 0 0 5px rgba(85, 204, 141, 0.6);
-}
-
-.inputField.activated {
-  border-color: var(--color-accent);
-}
-
-.error {
-  color: red;
-  font-size: 12px;
-  margin-top: 5px;
-  margin-bottom: 5px;
-  text-align: start;
-}
-
-button {
-  width: 100%;
-  height: 50px;
-  background-color: var(--color-accent);
-  border: none;
-  border-radius: 30px;
-  cursor: pointer;
-  margin-top: 16px;
-  font-size: 13px;
-  line-height: 18px;
-  font-weight: 700;
-  color: var(--color-main);
-  transition: background-color 0.3s ease;
-}
-
-button:hover {
-  background-color: #85cc8d;
-}
-
-button:disabled {
-  background-color: #ddd;
-}
-
-button[type="button"] {
-  background-color: #ccc;
-}
-
-.passwordWrapper {
-  position: relative;
-  display: flex;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-}
-
-.togglePassword {
-  position: absolute;
-  right: 16px;
-  top: 50%;
-  transform: translateY(-120%);
-  background: none! important;
-  border: none;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 32px;
-  height: 32px;
-  padding: 0;
-  opacity: 0.5;
-}
-
-.icon {
-  width: 20px;
-  height: 20px;
-  fill: none;
-  stroke: var(--color-main);
-  stroke-width: 1.8px;
-  transition: stroke 0.3s ease;
-                                                                                         
-}
-
-.icon:hover {
-  fill: var(--color-accent);
-}
-
-@media (min-width: 768px) {
-  .modalContent {
-      width: 600px;
-      padding: 25px;
-      border-radius: 25px;
-  }
-
-  .modalHeader {
-      font-size: 22px;
-  }
-
-  .modalText {
-      font-size: 17px;
-  }
-
-  .inputField {
-      height: 54px;
-      font-size: 15px;
-      padding: 14px;
-  }
-
-  button {
-      height: 55px;
-      font-size: 15px;
-  }
-}
-
-@media (min-width: 1440px) {
-  .modalContent {
-      width: 800px;
-      padding: 40px;
-      border-radius: 30px;
-  }
-
-  .modalHeader {
-      font-size: 24px;
-  }
-
-  .modalText {
-      font-size: 18px;
-  }
-
-  .inputField {
-      height: 50px;
-      font-size: 16px;
-      padding: 18px;
-  }
-
-  button {
-      height: 60px;
-      font-size: 16px;
-  }
-}
