+++ src/components/UI/SignInForm/SignInForm.module.css
@@ -327,12 +327,12 @@ a.signupLinkText {
     }
 
     a.signupLink {
-        font-size: 16px;
+        font-size: 14px;
         line-height: 24px;
     }
 
     .forgotPassword {
-        font-size: 16px;
+        font-size: 14px;
         line-height: 24px;
     }
 }
