+++ src/pages/ResetPasswordPage/ResetPasswordPage.module.css
@@ -1,0 +1,246 @@
+.section {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    padding: 24px 16px 24px 16px;
+    width: 343px;
+    height: 95%;
+    background-color: var(--color-gray);
+    border-radius: 30px;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.05);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.ResetPasswordSection {
+    width: 311px;
+    margin: 0;
+}
+
+h2.title {
+    font-size: 32px; 
+    font-weight: 700;
+    text-align: start;
+    color: var(--color-main);
+    margin-bottom: 32px;
+}
+
+input {
+    width: 100%;
+    height: 56px;
+    padding: 16px;
+    border: 1px solid rgba(47, 47, 47, 0.2);
+    border-radius: 15px;
+    font-size: 14px;
+    line-height: 18px;
+    background-color: var(--color-light);
+    transition: border-color 0.3s ease, box-shadow 0.3s ease;
+    margin-bottom: 16px;
+    padding-right: 48px;
+    color: inherit;
+}
+
+input.inactive {
+    border-color: rgba(47, 47, 47, 0.2); 
+    background-color: var(--color-light);
+}
+
+input:focus {
+    outline: none;
+    border-color: var(--color-accent);
+    box-shadow: 0 0 5px rgba(85, 204, 141, 0.6);
+}
+
+input.activated {
+    border-color: var(--color-accent);
+}
+
+.ErrorMessage {
+    color: red !important;
+    font-size: 14px;
+    margin-bottom: 10px;
+}
+
+input.errorInput {
+    border-color: red; 
+    background-color: rgba(255, 0, 0, 0.1); 
+}
+  
+  button {
+    width: 100%;
+    height: 50px;
+    background-color: var(--color-accent);
+    border: none;
+    border-radius: 30px;
+    cursor: pointer;
+    margin-top: 16px;
+    font-size: 13px;
+    line-height: 18px;
+    font-weight: 700;
+    color: var(--color-main);
+    transition: background-color 0.3s ease;
+  }
+  
+  button:hover {
+    background-color: #85cc8d;
+  }
+  
+  button:disabled {
+    background-color: #ddd;
+  }
+  
+  button[type="button"] {
+    background-color: #ccc;
+  }
+  
+  .passwordWrapper {
+    position: relative;
+    display: flex;
+    align-items: center;
+    width: 100%;
+    height: 100%;
+  }
+  
+  .togglePassword {
+    position: absolute;
+    right: 16px;
+    top: 50%;
+    transform: translateY(-120%);
+    background: none! important;
+    border: none;
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 32px;
+    height: 32px;
+    padding: 0;
+    opacity: 0.5;
+  }
+  
+  .icon {
+    width: 20px;
+    height: 20px;
+    fill: none;
+    stroke: var(--color-main);
+    stroke-width: 1.8px;
+    transition: stroke 0.3s ease;
+                                                                                           
+  }
+  
+  .icon:hover {
+    fill: var(--color-accent);
+  }
+  
+  @media (min-width: 320px) and (max-width: 375px) {
+    .section {
+        padding: 24px 16px 24px 16px;
+        width: 288px;
+        height: 95%;
+    }
+
+    .ResetPasswordSection {
+        width: 100%;
+        margin: 0;
+    }
+
+    h2.title {
+        font-size: 24px;
+        margin-bottom: 24px;
+    }
+
+    input {
+        font-size: 14px;
+        line-height: 20px;
+        height: 48px;
+        padding: 14px;
+    }
+
+    .button {
+        height: 50px;
+        font-size: 14px;
+        line-height: 18px;
+    }
+
+    .passwordWrapper {
+        position: relative;
+        display: flex;
+        align-items: center;
+        width: 100%;
+        height: 100%;
+    }
+
+    .togglePassword {
+        position: absolute;
+        right: 16px;
+        top: 50%;
+        transform: translateY(-120%);
+        background: none;
+        border: none;
+        cursor: pointer;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        width: 32px; 
+        height: 32px;
+        padding: 0;
+        opacity: 0.5;
+    }
+
+    .icon {
+        width: 20px;
+        height: 20px;
+    }
+
+    .errorMessage {
+        font-size: 12px;
+    }
+
+  }
+
+  @media (min-width: 768px) {
+     .section {
+        padding: 32px 134px 32px 134px;
+        width: 704px;
+        height: 95%;
+    }
+
+    .ResetPasswordSection {
+        width: 436px;
+        margin: 0;
+    }
+  
+    h2.title {
+        font-size: 36px;
+    }
+
+    input {
+        font-size: 16px;
+        line-height: 24px;
+    }
+
+    button {
+        height: 55px;
+        font-size: 15px;
+    }
+  }
+  
+  @media (min-width: 1440px) {
+    
+    .section {
+        padding: 32px 118px 32px 118px;
+        width: 672px;
+        height: 95%;
+    } 
+
+    .ResetPasswordSection {
+        margin: 0;
+    }
+
+    button {
+        height: 60px;
+        font-size: 16px;
+    }
+  }
